requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pandas>=1.3.3
numpy>=1.21.2
scikit-learn==0.24.2
nltk==3.6.3
matplotlib==3.4.3
seaborn==0.11.2
streamlit==0.84.2
plotly==5.3.1
faiss-cpu==1.7.2
whoosh==2.7.4
pydantic==1.8.2
node2vec==0.3.3
networkx==2.6.3
shap==0.40.0
requests>=2.28.0
python-multipart==0.0.5
aiofiles==0.7.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-dotenv==0.19.0
transformers==4.21.0
torch==1.12.1
textblob==0.17.1
wandb
